+++ R/esquisserUI.R
@@ -1,9 +1,12 @@
+
 #' @title Esquisse Shiny module
 #' 
 #' @description Launch \code{esquisse} in a classic Shiny app.
 #'
 #' @param id Module's id.
 #' @param header Logical. Display or not \code{esquisse} header.
+#' @param container Container in which display the addin, 
+#'  default is to use \code{esquisseContainer}, see examples.
 #' @param choose_data Logical. Display or not the button to choose data.
 #' @param insert_code Logical, Display or not a button to isert the ggplot
 #'  code in the current user script (work only in RStudio).
@@ -27,7 +30,6 @@
 #' @importFrom shinyWidgets prettyToggle
 #'
 #' @examples 
-#' 
 #' if (interactive()) {
 #' 
 #' 
@@ -39,22 +41,17 @@
 #' ui <- fluidPage(
 #'   tags$h1("Use esquisse as a Shiny module"),
 #'   
-#'   # Force scroll bar to appear (otherwise hidden by esquisse)
-#'   tags$style("html, body {overflow: visible !important;"),
-#'   
 #'   radioButtons(
 #'     inputId = "data", 
 #'     label = "Data to use:", 
 #'     choices = c("iris", "mtcars"),
 #'     inline = TRUE
 #'   ),
-#'   tags$div(
-#'     style = "height: 700px;", # needs to be in fixed height container
-#'     esquisserUI(
-#'       id = "esquisse", 
-#'       header = FALSE, # dont display gadget title
-#'       choose_data = FALSE # dont display button to change data
-#'     )
+#'   esquisserUI(
+#'     id = "esquisse", 
+#'     header = FALSE, # dont display gadget title
+#'     choose_data = FALSE, # dont display button to change data,
+#'     container = esquisseContainer(height = "700px")
 #'   )
 #' )
 #' 
@@ -95,9 +92,9 @@
 #' 
 #' 
 #' ui <- fluidPage(
-#'   tags$div( # needs to be in fixed height container
-#'     style = "position: fixed; top: 0; bottom: 0; right: 0; left: 0;", 
-#'     esquisserUI(id = "esquisse")
+#'   esquisserUI(
+#'     id = "esquisse", 
+#'     container = esquisseContainer(fixed = TRUE)
 #'   )
 #' )
 #' 
@@ -111,7 +108,10 @@
 #' 
 #' }
 #' 
-esquisserUI <- function(id, header = TRUE, choose_data = TRUE, insert_code = FALSE) {
+esquisserUI <- function(id, header = TRUE,
+                        container = esquisseContainer(),
+                        choose_data = TRUE, 
+                        insert_code = FALSE) {
   
   ns <- NS(id)
   
@@ -123,15 +123,15 @@ esquisserUI <- function(id, header = TRUE, choose_data = TRUE, insert_code = FAL
       class = "pull-right",
       miniTitleBarButton(inputId = ns("close"), label = "Close")
     ),
-    tags$div(
-      class = "pull-left",
-      if (isTRUE(choose_data) & isTRUE(header)) chooseDataUI(id = ns("choose-data"), class = "btn-sm")
-    )
+    if (isTRUE(choose_data) & isTRUE(header)) {
+      tags$div(
+        class = "pull-left",
+        chooseDataUI(id = ns("choose-data"), class = "btn-sm")
+      )
+    }
   )
     
-
-  ### addin
-  miniPage(
+  addin <- miniPage(
 
     # style sheet
     singleton(x = tagList(
@@ -190,4 +190,47 @@ esquisserUI <- function(id, header = TRUE, choose_data = TRUE, insert_code = FAL
     chartControlsUI(id = ns("controls"), insert_code = insert_code)
   )
 
+  if (is.function(container)) {
+    addin <- container(addin)
+  }
+  return(addin)
 }
+
+#' @param width,height The width and height of the container, e.g. \code{'400px'},
+#'  or \code{'100\%'}; see \code{\link[htmltools]{validateCssUnit}}.
+#' @param fixed Use a fixed container, e.g. to use use esquisse full page.
+#' 
+#' @rdname module-esquisse
+#' @export
+esquisseContainer <- function(width = "100%", height = "700px", fixed = FALSE) {
+  function(...) {
+    if (identical(fixed, FALSE)) {
+      tag <- tags$div(
+        style = sprintf("width: %s;", validateCssUnit(width)),
+        style = sprintf("height: %s;", validateCssUnit(height)),
+        ...
+      )
+    } else {
+      if (identical(fixed, TRUE)) {
+        tag <- tags$div(
+          style = "position: fixed; top: 0; bottom: 0; right: 0; left: 0;",
+          ...
+        )
+      } else if (is.numeric(fixed) & length(fixed) == 4) {
+        tag <- tags$div(
+          style = sprintf(
+            "position: fixed; top: %s; right: %s; bottom: %s; left: %s;",
+            fixed[1], fixed[2], fixed[3], fixed[4],
+          ),
+          ...
+        )
+      }
+    }
+    tagList(
+      singleton(tags$head(
+        tags$style("html, body {overflow: visible !important;")
+      )), tag
+    )
+  }
+}
+